fix(clippy): cast_possible_truncation across workspace (part 2)
Complete the lint with try_from or lint-clean rewrites at 12 more sites
(capi candidates/e2e tests, data content/initials/interp, datagen
system, store backend bench). Values are unchanged at every reachable
site: GUint/c_int FFI conversions saturate only past widths no
candidate count or buffer offset can reach (house precedent from the
iterators count conversions); slot_shift now runs its bound math in
usize with lossless widenings; the datagen bigram overflow keeps its
exact error message through try_from().map_err; the bench hash lanes
mask before converting, which is the same value the truncating cast
produced.

Five sites are deliberately left as reported STOPs: the float->u32
truncation in the pinned amplified-frequency law (engine session,
capi predict, and its two pinned mirrors) has no lossless std
conversion — std provides no TryFrom for float->int — and re-rounding
would move the frozen probe values; content.rs's header data_size
wrapping computation is u32 format semantics where the truncation is
load-bearing; and segment trellis's f64->f32 narrow reproduces
upstream's double rounding. Silencing lints is forbidden, so these
stay and are reported.

+++ crates/oxpinyin-data/src/initials.rs
@@ -202,11 +202,13 @@ impl InitialAlphabet {
 }
 
 /// Bit shift of slot `position`'s field: slot 0 is the most significant.
-const fn slot_shift(position: usize) -> u32 {
+const fn slot_shift(position: usize) -> usize {
     // Callers cap `position` at 24 (pack refuses slot >= 25, unpack walks
-    // 0..25); this pins the bound the subtraction relies on.
-    debug_assert!(SLOT_BITS * (position as u32 + 1) <= u128::BITS);
-    u128::BITS - SLOT_BITS * (position as u32 + 1)
+    // 0..25); this pins the bound the subtraction relies on. The math runs
+    // in `usize` so no `position` conversion is needed; both widening casts
+    // are lossless.
+    debug_assert!(SLOT_BITS as usize * (position + 1) <= u128::BITS as usize);
+    u128::BITS as usize - SLOT_BITS as usize * (position + 1)
 }
 
 /// Appends one slot's projection: `0` for the vowel sentinel, the key for
